+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * [HexaPDF::Content::CanvasComposer] for using high-level layout functionality
   on a single canvas
 * [HexaPDF::Content::Canvas#composer] for easily creating a canvas composer
+* [HexaPDF::Font::TrueTypeWrapper#bold?] and [HexaPDF::Font::Type1Wrapper#bold?]
+  for determining whether a font is bold
 
 ### Changed